delete dtype

+++ build.py
@@ -42,7 +42,7 @@ def compile(name, embed_mat, seq_len, class_num):
     vocab_num, embed_len = embed_mat.shape
     embed = Embedding(input_dim=vocab_num, output_dim=embed_len,
                       weights=[embed_mat], input_length=seq_len, trainable=True)
-    input = Input(shape=(seq_len,), dtype='int32')
+    input = Input(shape=(seq_len,))
     embed_input = embed(input)
     func = map_item(name, funcs)
     output = func(embed_input, class_num)
